fix(商品模块-获取商品列表): 销量为0时返回null的问题

+++ src/main/java/com/shopping/shoppingApi/vo/IndexProductVO.java
@@ -28,7 +28,7 @@ public class IndexProductVO implements Serializable {
     @Schema(description = "是否热门")
     private Boolean isHot;
     @Schema(description = "总销量")
-    private BigDecimal totalSaleAmount;
+    private Integer totalSaleAmount;
     @Schema(description = "是否新品")
     private Boolean isNew;
     @Schema(description = "是否促销")
